requirements.txt
ultralytics
opencv-python
numpy
pygame
pyserial
python-dotenv
twilio
firebase-admin
google-cloud-storage
